+++ jerrors.go
@@ -1,0 +1,32 @@
+package main
+
+// Jerror interface for JSON errors
+type Jerror interface {
+	Error() string
+	Code() int
+}
+
+// JSONerror enables returning enriched errors with JSON status
+type JSONerror struct {
+	error string
+	code  int
+}
+
+// Error() provide the classical error status
+func (j JSONerror) Error() string {
+	return j.error
+}
+
+// Code provides JSON error code to be returned
+func (j JSONerror) Code() int {
+	return j.code
+}
+
+// newJSONerror return JSON error from string
+func newJSONerror(code int, e string) JSONerror {
+	return JSONerror{error: e, code: code}
+}
+
+func errorToJSON(code int, e error) JSONerror {
+	return JSONerror{error: e.Error(), code: code}
+}
